refactor: rename annotator_model.py to model.py
With the legacy CLIP model.py removed in the previous commit, the
canonical CellTypeAnnotator can reclaim the obvious filename. Updates
the two import sites (predict.py, tests/test_canonical_inference.py)
to match. This also lines the import path up with the training repo
(deepcelltypes-cell-type-assignment-pytorch), which has been using
`deepcelltypes.model.CellTypeAnnotator` all along — easing the
upcoming training-pipeline merge.
