+++ deepcell_types/model.py
