+++ tests/test_canonical_inference.py
@@ -6,7 +6,7 @@ import torch
 from torch.utils.data import DataLoader
 from numcodecs import Zstd
 
-from deepcell_types.annotator_model import create_model
+from deepcell_types.model import create_model
 from deepcell_types.dataset import PatchDataset
 from deepcell_types.dct_kit.config import DCTConfig
 from deepcell_types.predict import (
